Make find_LGR_Image(..) raise ValueError
When find_LGR_Image does not find what it is looking for, it used to
return False. False and 0 is the same in python, which means that it was
impossible to know if if an image was found at index zero, or no image
was found at all (without doing further checks). Therefore, we raise a
ValueError instead when we do not find anything, which mirrors the
behavior of the built-in list.index(..).

This also adds a test for the "did not find" case of find_LGR_Image(..).