Fix and rename incorrect GroundTouchBEvent to AppleTouchEvent
There is only one type of ground touch event in Elma. The event type that
has been previously interpreted as GroundTouchBEvent is actually an event
that is generated when touching an apple. Object touch event is also always
generated together with the apple touch event and presumably the only reason
why apple touch event exists is just for sound generation.