Simplify public API
This commit simplifies the public API so that most classes and functions
can be imported directly from the main package instead of submodules.
The public API is also decoupled from the internal code structure to
make it easier to change code without breaking the public API in future.
Also __all__ has been added to most files to allow safer import *
just in case.